requirements.txt
httpx>=0.27
loguru>0.7
orjson>=3.10
spacy>=3.8
soundfile>=0.12
pyrubberband>=2.2
loguru>0.7
anyio[trio]>=4.6
ffmpeg-python>=0.2.0